+++ src/main/java/io/apiman/cli/core/common/action/ModelListAction.java
@@ -1,6 +1,22 @@
-package io.apiman.cli.api.action.common;
+/*
+ * Copyright 2016 Pete Cornish
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-import io.apiman.cli.api.exception.ActionException;
+package io.apiman.cli.core.common.action;
+
+import io.apiman.cli.exception.ActionException;
 import io.apiman.cli.util.JsonUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -14,7 +30,7 @@ import java.util.List;
 import static io.apiman.cli.util.LogUtil.OUTPUT;
 
 /**
- * @author Pete
+ * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
 public abstract class ModelListAction<M, A> extends AbstractModelAction<M, A> {
     private static final Logger LOGGER = LogManager.getLogger(ModelListAction.class);