+++ src/main/java/io/apiman/cli/core/gateway/action/AbstractGatewayCreateAction.java
@@ -1,10 +1,26 @@
+/*
+ * Copyright 2016 Pete Cornish
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.apiman.cli.core.gateway.action;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import io.apiman.cli.api.action.common.ModelCreateAction;
+import io.apiman.cli.core.common.action.ModelCreateAction;
 import io.apiman.cli.core.gateway.GatewayApi;
 import io.apiman.cli.core.gateway.GatewayMixin;
-import io.apiman.cli.api.exception.ActionException;
+import io.apiman.cli.exception.ActionException;
 import io.apiman.cli.core.gateway.model.Gateway;
 import io.apiman.cli.core.gateway.model.GatewayConfig;
 import io.apiman.cli.core.gateway.model.GatewayType;
@@ -12,7 +28,9 @@ import io.apiman.cli.util.JsonUtil;
 import org.kohsuke.args4j.Option;
 
 /**
- * @author Pete
+ * Shared functionality for actions requiring a {@link Gateway}.
+ *
+ * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
 public abstract class AbstractGatewayCreateAction extends ModelCreateAction<Gateway, GatewayApi>
         implements GatewayMixin {