+++ src/main/java/io/apiman/cli/core/gateway/action/GatewayTestAction.java
@@ -1,7 +1,23 @@
+/*
+ * Copyright 2016 Pete Cornish
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.apiman.cli.core.gateway.action;
 
-import io.apiman.cli.api.exception.ActionException;
-import io.apiman.cli.api.exception.ExitWithCodeException;
+import io.apiman.cli.exception.ActionException;
+import io.apiman.cli.exception.ExitWithCodeException;
 import io.apiman.cli.core.gateway.GatewayApi;
 import io.apiman.cli.core.gateway.model.GatewayTestResponse;
 import io.apiman.cli.util.JsonUtil;
@@ -15,7 +31,9 @@ import java.text.MessageFormat;
 import static io.apiman.cli.util.LogUtil.OUTPUT;
 
 /**
- * @author Pete
+ * Test a gateway.
+ *
+ * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
 public class GatewayTestAction extends AbstractGatewayCreateAction {
     private static final Logger LOGGER = LogManager.getLogger(GatewayTestAction.class);