+++ src/main/java/io/apiman/cli/util/ApiUtil.java
@@ -1,7 +1,23 @@
+/*
+ * Copyright 2016 Pete Cornish
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.apiman.cli.util;
 
 import com.google.common.io.CharStreams;
-import io.apiman.cli.api.exception.ActionException;
+import io.apiman.cli.exception.ActionException;
 import retrofit.RetrofitError;
 import retrofit.client.Response;
 
@@ -12,7 +28,9 @@ import java.net.HttpURLConnection;
 import java.util.function.Supplier;
 
 /**
- * @author Pete
+ * Shared API utility methods.
+ *
+ * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
 public class ApiUtil {
     public static void invokeAndCheckResponse(Supplier<Response> request) throws ActionException {